Move domain base command
